+++ src/main/java/rw/ehealth/service/user/IDepartemtService.java
@@ -12,5 +12,5 @@ public interface IDepartemtService {
 	Department findPerName(String name);
 	List<Department> findPerHospital(Hospital hospital);
 
-	Department findPerId(int id);
+	Department findPerId(Long id);
 }
